Webook/requirements.txt
asgiref==3.8.1
cassandra-driver==3.29.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.7
django-cassandra-engine==1.9.0
Faker==26.2.0
geomet==0.2.1.post1
idna==3.7
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
scylla-driver==3.26.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
